requirements.txt
apache-airflow==1.10.15
boto3==1.20.26
